Fix clippy 1.84 lints - including small API changes
The API change is that various process_{capture,render}_frame() methods now take `&mut [Vec<f32>]` instead of `&mut Vec<Vec<f32>>`.

Such a change should be compatible, as every `&mut Vec<_>` is also a `&mut [_]`, but strictly-speaking it may be a semver-breaking change? Maybe we shouldn't care that much.